Employe_Monitoring_System/background_script/requirements.txt
pyautogui
pillow
